refactor: Package, class structure and names for clarity
