+++ src/main/java/baseball/domain/BaseballNumber.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import java.util.Collections;
 import java.util.List;
