+++ src/main/java/baseball/domain/Constants.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public final class Constants {
     public static final int MAX_DIGITS = 3;
